Allow #sitting on eggs.
If you are polymorphed into a female monster of appropriate type,
then #sitting on an egg will cause it to hatch faster.  If the egg
is yours, and it will never hatch (due to having been kept too long
in a container, for example), then you will "feel sad for a moment".