Allow heavy creatures to shatter chests (and eggs) by sitting on them.
This comes from a YANI by Mark Rae. Creatures with weight of 2600
or greater (e.g. dragons, titanotheres) have an 80% chance of
"completely destroying" any box or chest they #sit on.  Fragile
objects inside the box have a chance of destruction as well, just
as if the box had been destroyed by #forcing.
http://groups.google.com/group/rec.games.roguelike.nethack/msg/b9e08d21fa4645d2

A large creature #sitting on an egg of inappropriate type will
cause the egg to splat. Splatting a cockatrice egg in this manner
does NOT cause stoning, for the same reason that #sitting on a
cockatrice corpse does not cause stoning.
http://nethack.wikia.com/wiki/Inventory#Pants