Fix "over-cautiousness" when a cockatrice-wielding golem removes its gloves.
If you are wearing gloves and wielding a cockatrice corpse, and then
you take off your gloves, you will be insta-petrified. If you survive
(e.g. thanks to an AoLS), the corpse will be unwielded. This is
unnecessary in the case that you are a flesh golem (removing your gloves
turns you into a stone golem), or in the case that you are a stone golem
already.