Allow any user to enter wizard mode via "nethack -D".
It's silly to forbid regular users from using wizard mode; it's even
sillier to have the guarding condition be the actual username "wizard".
This change enables "nethack -D" for all users.

The real reason for this change is to make it easier for me to test
my changes in wizard mode.