Report unidentified gemstone rings as, e.g., "green stone rings".
This comes from a YANI by Joe Clark. He pointed out that it
doesn't make sense to recognize "an emerald ring" if you
haven't yet identified emerald gems.
http://groups.google.com/group/rec.games.roguelike.nethack/msg/d00145f38d7c0c7e
http://www.club.cc.cmu.edu/~ajo/free-software/nethack/unidring343.diff

Vice versa, as soon as you identify a ring, the corresponding
gem will also be identified. One thing to notice is that the
wizard role starts with two identified rings, which means that
wizards have a decent chance of being able to identify certain
gems without needing to acquire a touchstone or blessed ?oI.

Wishing for "a green stone ring" or "a white stone ring" will
now give you either emerald or jade, resp. opal or diamond,
with 50/50 probability.

The gemstones affected by this patch are:

 - ruby (red stone)
 - agate (orange stone)
 - topaz (yellowish brown stone)
 - emerald (green stone)
 - jade (green stone)
 - sapphire (blue stone)
 - opal (white stone)
 - diamond (white stone)