Fix C343-100: crash if thrown/kicked potion hits iron bars and breaks.
This bugfix matches the one in Patric Mueller's UnNetHack, except
that I also fix the similar crash in kick_object(). That crash is
100 times less visible, though, since kicking a potion (or other
breakable object, such as glass) has a 99% chance of breaking it
even before it starts moving.

Like C343-179, this crash happens when a callee breaks and destroys
an object, and then the caller attempts to refer to the original
object (via a pointer that is now dangling).