Fix(?) C343-19: apparent bug when dipping greased acid in a fountain.
The official bug description is "Dipping acid in a fountain may
not destroy the acid."  This ought to mean that dipping acid in
a fountain *should* always destroy the acid, but in fact does not
(i.e., the DevTeam's list always describes the buggy behavior
rather than the intended fix).  However, in the current 3.4.3 code,
it's possible both for acid to survive "boiling vigorously", and
for acid to mysteriously disappear after being "protected" by a
layer of grease.

Old behavior:
    - #dipping greased acid in a fountain or pool gives
      "Your potion is protected by the layer of grease!",
      but the acid never survives anyway. Clearly a bug.
    - #dipping greased acid in a potion of water gives
      "Your potion is protected by the layer of grease!";
      the acid and water both survive.
    - #dipping ungreased acid in a potion of water always
      "boils vigorously"; the acid always survives (but the
      potion of water never does).
    - #dipping ungreased acid in a fountain always "boils
      vigorously"; the acid survives 50% of the time.
      Probably a bug.
    - #dipping ungreased acid in a pool always "boils
      vigorously"; the acid never survives.

The behavior when dipping greased acid in a fountain/pool
is clearly a bug.

I don't like the apparently intentional difference in
survival rates between dipping in a fountain/pool versus a
potion of water. The intent seems to be that dipping acid
into a large body of water (fountain/pool) causes it to
be used up, but dipping acid into a relatively small bottle
of water doesn't cause such a big reaction, so the potion
survives. I don't think there's any scientific justification
for this behavior, but I'll strive to preserve it.

With this patch, the new behavior is:
    - Greased acid always survives #dipping.
    - #dipping ungreased acid in a potion of water always
      "boils vigorously"; the acid always survives (but the
      potion of water never does). Same as old behavior.
    - #dipping ungreased acid in a fountain or pool
      always "boils vigorously"; the acid never survives.

So this patch fixes one obvious bug (the behavior of
greased acid), and also brings the behavior of fountains
into line with the existing behavior of pools by causing
acid to get used up more often. I'm pretty sure that this
is at least close to what the DevTeam had in mind.