Fix C343-364: misnaming an artifact never picks 'z'.
If the player's hand slips while naming an object (for example,
if he's trying to name his warhammer "Mjollnir", or if he's trying
to name a stack of elven daggers "Sting", or if Sting already
exists), then the last letter of the engraving will be changed
to a random letter between 'a' and 'z'. But the code implementing
this had forgotten the behavior of rn2(): the upper endpoint of
the range is never chosen. So really the random letter was always
between 'a' and 'y'. This patch corrects the error.

I'm also adding a comma to the message, since I think it reads
much better that way.