Once you're in a pit, allow "<" to try to get out of it.
The horizontal directions already work to get out of a pit (if you
keep trying long enough, that is). It's even more logical to try
to get out of the pit by moving "up", but for some reason that was
never implemented.

At the same time, I've reordered the checks so that it is newly
possible for you to enter a pit (via ">") while your steed is
eating, and impossible for you to "ease yourself into" a pit
while engulfed. Personally, I don't think it makes any sense that
you can float_up/float_down while held, engulfed, or swallowed,
either; but I've left that behavior intact since it's clearly
intentional.

This patch and the next one, taken together, address C343-16
and C343-17.