Fix C343-16 and C343-17: bad interactions between xorns and pits.
C343-16: "Player polymorphed as a xorn can't pick up items in pits."
C343-17: "After polymorphing into a xorn and returning to human form,
player can be stuck in a pit that isn't there."

Both issues are due to the DevTeam's questionable decision to make
xorns (and other "Passes_walls" monsters such as earth elementals)
absolutely immune to pit traps. In the old code, being poly'ed into
a xorn automatically got you out of any pit you were in, and if you
fell into another one, you immediately (and silently) got out of
that one, too. But the code was poorly done: it merely set u.utrap
to 0, and then upon returning to human form it set u.utrap back to
rn1(6,2) if u.utraptype was still TT_PIT. That was doubly incorrect.
First, a xorn player could see "You fall into your pit!" and then
actually *not* be in the pit (due to xorns' get-out-of-jail-free
status); this resulted in the symptom observed in C343-16. Second,
a player in a pit could polymorph into a xorn (getting out of jail
free), move a couple of spaces away, and poly back to a human; since
the get-out-of-jail-free code didn't clear u.utraptype, you'd
wind up trapped in a pit that wasn't there: C343-17.

This new code replaces all that with the simple rule that xorns
can be pitted just like anybody else. However, when a xorn attempts
to leave a pit by moving horizontally, he'll always succeed --- and
in fact won't even lose that turn. A xorn trying to move "<" out of
a pit suffers the same consequences as anyone else, of course.