Fix C343-406: spelling error in one of the coyote names.
The list is taken from spellings given in actual Warner Bros.
cartoons, so *apparent* inconsistencies such as "Famishus" versus
"Famishius" are actually intentional. There are a few more names
given in the cartoons that aren't listed here, according to
Wikipedia:

    - Desertous-operativi Idioticus (1962)
    - Dogius Ignoramii (1994)
    - Poor Schnookius (2000)
    - Desertus Operatus Idioticus (2003)

For what it's worth, the current coynames[] array was put
together in 1999. That doesn't antedate all of the missing
cartoons, but definitely antedates the Wikipedia list.