Fix C343-63: #dipping/quaffing from a fountain while swallowed.
As I remarked in commit 97e1a2f1, I think it's odd that you're
able to interact with the Z-axis while swallowed. ("You float down,
but you are still swallowed.") But C343-63 clarifies that you're
*definitely* not supposed to be able to interact with dungeon
features such as fountains and sinks.

After this fix, neither (q)uaffing nor #dipping offer to interact
with fountains or sinks while you are swallowed.