Fix C343-30: stat exploit using cursed scroll of destroy armor.
The official bug description is "Cursed scroll of destroy armor
on cursed armor doesn't always interact correctly." This is
subtle; the problem is that two armor items --- helm of brilliance
and gauntlets of dexterity --- actually give a stats benefit
based on their current enchantment. The code for ?oEA correctly
calls adj_abon() so that increases in o->spe are reflected in
the hero's stats, but the corresponding codepath in ?oDA simply
decremented o->spe without also reducing the hero's stats. The
result: repeatedly enchanting and cursed-destroying a cursed HoB
could infinitely increase your Int/Wis up to the maximum, and
likewise for gauntlets of dexterity resp. Dex.

This fix simply adds the missing call to adj_abon().