Fix C343-104: bad grammar from angry shopkeeper giving a price.
Shopkeepers address you differently when they are hostile.
Old code:

  - "For you, esteemed sir; only 30 for this food ration."
  - "For you, scum ; only 40 for this food ration."

New code:

  - "For you, esteemed sir, only 30 for this food ration."
  - "For you, scum, 40 for this food ration."

The extra space in the angry message was clearly a mistake.
I think the semicolons were a grammatical error as well. And
lastly, I don't think an angry shopkeeper would still be trying
to coax you with low prices, so I've removed the word "only" in
the angry case.