Fix SC343-8, C343-94: bugs when re-animating a statue containing gold.
C343-94: "Probing a re-animated statue may cause a panic."
SC343-8: "Re-animating a statue containing gold produces too much gold."

Both of these bug descriptions refer to the same issue. When you
stone a monster, the statue is just a plain old object, containing
other plain old objects. If the monster was carrying gold (mon->mgold),
then a gold object is created in the statue's inventory.

But if the stoned monster was unique, or tame, or a shopkeeper, then
NetHack attaches a saved representation of the original monster to
the statue object. This is so that reanimating the statue of a pet
will give you your original pet back, for example.

The first bug is that this saved representation (save_mtraits(),
montraits()) still has the original monster's "mgold", even though
the gold is also present as an object in the statue's contents.
The second bug is that, on reanimation, the statue's contents are
transferred back to the monster without a special case for !GOLDOBJ;
this means that the reanimated monster will have a "gold object" in
its inventory, which violates the !GOLDOBJ invariant.

The upshot is, (1), if you probe the reanimated monster, the game
will see a monster with mgold>0 *and* a gold object in its inventory,
and this will crash the game. And (2), if you kill the reanimated
monster, it will drop its existing gold object plus a newly created
gold object to account for its mgold; these will correctly merge when
dropped, so you'll get twice as much gold as you should. This bug can
be exploited to repeatedly double a quantity of gold.

My fix makes sure that save_mtraits() never saves a monster's mgold,
and also that animate_statue() correctly handles gold objects in the
statue's contents.

My fix is different from Patric Mueller's fix in UnNetHack. That fix
doesn't handle the !GOLDOBJ case in animate_statue(), yet apparently
does fix the issue; I'm not sure why.